requirements.txt
ai21>=3.1.0
faiss-cpu>=1.7.4
langchain>=0.1.0
pydantic>=2.5.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
chroma-hnswlib>=0.7.3
chromadb>=0.4.18
